OOPL Assignment 1
Define a Class to represent a Bank account which include the following members such as name of depositer,account no., withdrawal account. Some additional balance account also include the member function to asign initial value to deposite and to withdraw an account, after and before checking balance to display name and balance.