OOPL Assignment 2
A book shop maintains the inventary of books that are being sold at the shop. The list includes details such as author,title,price,publisher and stock position. Whenever a customer wants a book, the sales person inputs the title and author and the system searches the list and displays whether it is available or not. If it is not available or not . If it is not , an appropriate message is displayed. If it is then the system displays the book details and requests for the number of copies required. If the requested are available, the total cost of the required copies is displayed; otherwise a message "Required copies not in Stock" is displayed. Design a system using class called book with suitable member functions and Construction. Use new operator in construction to allocate memory space required. Implement C++ program.